Inline selectors to components
