+++ packages/scrabble-solver/src/components/Board/Board.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable max-lines, max-statements */
 
 import { FloatingPortal, ReferenceType } from '@floating-ui/react';
+import { createSelector } from '@reduxjs/toolkit';
 import { EMPTY_CELL } from '@scrabble-solver/constants';
 import classNames from 'classnames';
 import { CSSProperties, FocusEventHandler, FunctionComponent, useCallback, useState } from 'react';
@@ -9,14 +10,16 @@ import { useDispatch } from 'react-redux';
 
 import { useAppLayout } from 'hooks';
 import { LOCALE_FEATURES } from 'i18n';
+import { findCell } from 'lib';
 import { TRANSITION } from 'parameters';
 import {
   boardSlice,
   cellFiltersSlice,
+  selectBoard,
   selectCellFilters,
   selectInputMode,
   selectLocale,
-  selectRowsWithCandidate,
+  selectResultCandidateCells,
   selectShowCoordinates,
   solveSlice,
   useTypedSelector,
@@ -30,6 +33,10 @@ import { useBoardStyle, useFloatingActions, useFloatingFocus, useFloatingInputPr
 interface Props {
   className?: string;
 }
+
+const selectRowsWithCandidate = createSelector([selectBoard, selectResultCandidateCells], (board, cells) => {
+  return board.rows.map((row, y) => row.map((cell, x) => findCell(cells, x, y) || cell));
+});
 
 export const Board: FunctionComponent<Props> = ({ className }) => {
   const dispatch = useDispatch();
