+++ packages/scrabble-solver/src/hooks/useColumns.ts
@@ -1,6 +1,8 @@
+import { createSelector } from '@reduxjs/toolkit';
 import { useMemo } from 'react';
 
-import { selectColumns, useTypedSelector } from 'state';
+import { LOCALE_FEATURES } from 'i18n';
+import { selectLocale, selectShowCoordinates, useTypedSelector } from 'state';
 import { ResultColumnId } from 'types';
 
 import { useMediaQueries } from './useMediaQueries';
@@ -12,6 +14,31 @@ const COLUMNS_S = [...COLUMNS_XS, ResultColumnId.BlanksCount, ResultColumnId.Wor
 const COLUMNS_M = [...COLUMNS_XS];
 
 const COLUMNS_L = [...COLUMNS_XS];
+
+const selectColumns = createSelector([selectLocale, selectShowCoordinates], (locale, showCoordinates) => {
+  const { consonants, vowels } = LOCALE_FEATURES[locale];
+  const columns: ResultColumnId[] = [
+    ResultColumnId.Word,
+    ResultColumnId.TilesCount,
+    ResultColumnId.BlanksCount,
+    ResultColumnId.WordsCount,
+    ResultColumnId.Points,
+  ];
+
+  if (showCoordinates !== 'hidden') {
+    columns.push(ResultColumnId.Coordinates);
+  }
+
+  if (vowels) {
+    columns.push(ResultColumnId.VowelsCount);
+  }
+
+  if (consonants) {
+    columns.push(ResultColumnId.ConsonantsCount);
+  }
+
+  return columns;
+});
 
 export const useColumns = (): Partial<Record<ResultColumnId, boolean>> => {
   const columns = useTypedSelector(selectColumns);
