+++ packages/scrabble-solver/src/modals/RemainingTilesModal/RemainingTilesModal.tsx
@@ -1,10 +1,11 @@
+import { createSelector } from '@reduxjs/toolkit';
 import { FunctionComponent, memo } from 'react';
 
 import { Badge, Modal } from 'components';
 import { LOCALE_FEATURES } from 'i18n';
-import { getTileSizes } from 'lib';
+import { getRemainingTilesGroups, getTileSizes } from 'lib';
 import { REMAINING_TILES_TILE_SIZE } from 'parameters';
-import { selectLocale, selectRemainingTilesGroups, useTranslate, useTypedSelector } from 'state';
+import { selectLocale, selectRemainingTiles, useTranslate, useTypedSelector } from 'state';
 
 import { Character } from './components';
 import styles from './RemainingTilesModal.module.scss';
@@ -14,6 +15,8 @@ interface Props {
   isOpen: boolean;
   onClose: () => void;
 }
+
+const selectRemainingTilesGroups = createSelector([selectRemainingTiles], getRemainingTilesGroups);
 
 const RemainingTilesModalBase: FunctionComponent<Props> = ({ className, isOpen, onClose }) => {
   const translate = useTranslate();
